+++ lib/Crypto/SelfTest/Util/test_number.py
@@ -82,6 +82,24 @@ class MiscTests(unittest.TestCase):
         self.assertEqual(2, number.ceil_div(8, 4))
         self.assertEqual(3, number.ceil_div(9, 4))
 
+    def test_getPrime(self):
+        """Util.number.getPrime"""
+        self.assertRaises(ValueError, number.getPrime, -100)
+        self.assertRaises(ValueError, number.getPrime, 0)
+        self.assertRaises(ValueError, number.getPrime, 1)
+
+        bits = 4
+        for i in range(100):
+            x = number.getPrime(bits)
+            self.assertEqual(x >= (1 << bits - 1), 1)
+            self.assertEqual(x < (1 << bits), 1)
+
+        bits = 512
+        x = number.getPrime(bits)
+        self.assertNotEqual(x % 2, 0)
+        self.assertEqual(x >= (1 << bits - 1), 1)
+        self.assertEqual(x < (1 << bits), 1)
+
     def test_getStrongPrime(self):
         """Util.number.getStrongPrime"""
         self.assertRaises(ValueError, number.getStrongPrime, 256)